requirements.txt
customtkinter==5.2.2
tkcalendar==1.6.1
yfinance==0.2.54
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==5.1.0
matplotlib==3.9.2
plyer==2.1.0
python-dateutil==2.9.0.post0
